review fix & refactor
