+++ src/com/c1k4d4/tasks/week3/oop/Task1Config.java
@@ -1,0 +1,10 @@
+package com.c1k4d4.tasks.week3.oop;
+
+public class Task1Config {
+
+    static final int BUILDING_NUMBER = 47;
+    static final int STOREYS_AMOUNT = 5;
+    static final int APARTMENTS_AMOUNT = 3;
+    static final int ROOMS_AMOUNT = 3;
+
+}
